Add FizzBuzz implementation in Lisp
Created a new Lisp function to run the FizzBuzz program from 1 to 100. This implementation checks and replaces multiples of 3 with "Fizz", multiples of 5 with "Buzz" and multiples of both with "FizzBuzz". This not only provides a simple way to test Lisp's functionality, but also demonstrates the efficiency of Lisp's syntax and structure.